requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.5
django-crispy-forms==1.14.0
django-on-heroku==1.1.2
django-phonenumber-field==6.1.0
gunicorn==20.1.0
phonenumbers==8.12.50
Pillow==9.1.1
psycopg2-binary==2.9.3
py3dns==3.2.1
python-dotenv==0.20.0
sqlparse==0.4.2
validate-email-address==1
whitenoise==6.2.0
